% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageCropper.R
\name{imageCropper}
\alias{imageCropper}
\title{Image Cropper Function}
\usage{
imageCropper(width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{width}{Fixed width for widget (in css units). The default is NULL, which
results in intelligent automatic sizing based on the widget’s
container.}

\item{height}{Fixed height for widget (in css units). The default is NULL, which
results in intelligent automatic sizing based on the widget’s
container.}

\item{elementId}{Use an explicit element ID for the widget
 Useful if you have other JavaScript that needs to
 explicitly discover and interact with a specific widget instance.
in any other case leave as NULL which results in an
automatically generated one.}
}
\description{
This function maps images to a html component to complete the
             rendering of images...
}
\examples{
imageCropper()

}
\keyword{cropper.js}
